Initial OpenCL support infra and dev direction
This commit introduces the initial framework to support OpenCL. Parts of the
code can easily be shared with the CUDA implentation, but this is work for a
future commit. Keeping things separate will help with the transition and being
able to build both CUDA as well as OpenCL variants.

Because OpenCL setup is extensive (compared to CUDA), this part is contained
in opencl-platform directory and done only once. This does not remove the
initial OpenCL init overhead, and this is not avoidable but this does remove
the need to re-init/create/compile devices and kernels.

All kernel code are contained in cl files. This is an OpenCL limitation since
the distribution of the library should include the cl files hence this is a
problem. Future commits can embed the OpenCL into a string and have it already
in the binary/library. This should be done as a last step, once everything is
working and in place.

The following new objects compile:
debug/cl_aes_cbc.o
debug/cl_chacha_cbc.o
debug/cl_init_platform.o

The CUDA compilation path is not altered, nor are CUDA objects.